+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed code-formatting
 - Renamed `WsjcppCore::extractURLProtocol` to `wsjcpp::extract_url_protocol`
 - Fixed #38: Added test for extract_url_protocol
+- Renamed and redesign `WsjcppCore::recoursiveRemoveDir` to `wsjcpp::recursive_remove_dir`
 
 ## [v0.2.5] - 2026-05-19 (2026 May 19)